db-writer/requirements.txt
confluent-kafka
psycopg2-binary==2.9.9
